Delete redundant json files and create build.make for main
Deleted obsolete json files in the cmake-build-debug directory. Added new build.make file for the main executable. The build.make file contains a collection of rules for building the executable and an associated clean-up script. Changes improve the organization of the build-debug directory and specify build instructions for the main project file.