Delete code documentation HTML version
The HTML version of the code documentation was removed. This change aims to avoid redundancy and to keep the documentation focused mainly on the markdown files, which are more straightforward and easier to maintain. Going forward, any updates or changes to the documentation will only need to be made in the markdown files. This change helps simplify the project structure and reduce unnecessary bloat.