Refactor project documentation HTML styles
The styles for the project documentation HTML page were updated to improve visual hierarchy and readability. Styling for table of contents (#toc) was changed to ensure better visibility and accessibility. Adjustments were made to the #toc and #content layout, as well as list styles for table of contents and color tweaks for hyperlink states within the table