Update project files and class descriptions
This commit introduces improvements and additions to project files including the HTML documentation, README, and main.cpp. It also adds in-depth description and clarification for FileReader and HeaderMaker classes in main.cpp. More detailed file structures and features of the system were written in the markdown and HTML documentation. New functionalities such as mobile responsiveness for the Table of Contents in the HTML documentation were also implemented. The “FileReader” and “HeaderMaker” classes were updated. Name and contribution details of the contributors of the project were specified and added more comprehensively.