Add project documentation HTML file
This commit adds an HTML file "Documentacion del Proyecto.html" with the initial boilerplate code. The code contains the fundamental HTML structure, styling in form of a CSS embedded in the HTML file, and a Javascript functionality for a Dark Mode toggle. This documentation will serve as the project's primary reference and will be continuously updated throughout the development process.